refactor(types): extract chain-info module
